Modified total_editing.py to calculate codon prefs.
Codon preferences are calculated simply as the number of times each codon shows
up in the genomic and mRNA sequence. Sequence starts and stops as before
(compares only the aligned regions), and it is ensured that each codon is counted
only once by recording codons only when codon_pos is 3. Missing data at the
beginning of the mRNA sequence is compensated by ensuring that only full
codons are counted for the first codon recorded. The original SeqPair class
in classes.py was copied and slightly modified to yield CodonPair, which is
present in the same file.